+++ hibernate/springData/src/main/java/ru/alexkm07/spdata/repository/OrderRowRepository.java
@@ -3,5 +3,5 @@ package ru.alexkm07.spdata.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.alexkm07.spdata.model.OrderRow;
 
-public interface OrderRowRepository extends CrudRepository<OrderRow,Integer> {
+public interface OrderRowRepository extends CrudRepository<OrderRow,Long> {
 }
