+++ hibernate/springData/src/main/java/ru/alexkm07/spdata/service/OrderRowController.java
@@ -17,7 +17,7 @@ public class OrderRowController {
         orderRowRepository.save(orderRow);
     }
 
-    public OrderRow getById(Integer id){
+    public OrderRow getById(Long id){
         Optional<OrderRow> orderRow = orderRowRepository.findById(id);
         return orderRow.get();
     }
